pushing final project to github

+++ light-novel-logger/src/main/java/light/novel/logger/controller/model/LightNovelData.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class LightNovelData {
 	private Long lightNovelId;
-	private int volumeNumber;
+	private Long volumeNumber;
 	private int pageCount;
 	private String description;
